requirements.txt
# YOLOv8 People Detection and Tracking Requirements

# Core dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# Backend API dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Authentication dependencies
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1

# Export dependencies
pandas>=2.0.0
reportlab>=4.0.0

# WebSocket support (included with fastapi)
websockets>=12.0
